Blogs: Moved private static helper methods to the "BlogRssUtils" class for better reuse
