+++ OurUmbraco/Community/BlogPosts/BlogRssUtils.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Globalization;
+using System.Xml.Linq;
+using Skybrud.Essentials.Xml.Extensions;
 
 namespace OurUmbraco.Community.BlogPosts {
     
@@ -16,6 +18,55 @@ namespace OurUmbraco.Community.BlogPosts {
             // Parse a valid RFC 822 formatted date
             return DateTimeOffset.ParseExact(str, "ddd, dd MMM yyyy HH:mm:ss K", CultureInfo.InvariantCulture).ToLocalTime();
 
+        }
+        public static string GetFileExtension(string blogLogoUrl)
+        {
+            var extension = ".png";
+            var url = blogLogoUrl;
+            if (url.Contains("?"))
+                url = blogLogoUrl.Substring(0, blogLogoUrl.IndexOf("?", StringComparison.Ordinal));
+
+            var lastUrlSlug = url.Substring(url.LastIndexOf("/", StringComparison.Ordinal));
+            if (lastUrlSlug.Contains("."))
+                extension = lastUrlSlug.Substring(lastUrlSlug.LastIndexOf(".", StringComparison.Ordinal));
+            return extension;
+        }
+
+        public static string RemoveLeadingCharacters(string inString)
+        {
+            if (inString == null)
+                return null;
+
+            var substringStart = 0;
+            for (var i = substringStart; i < inString.Length; i++)
+            {
+                var character = inString[i];
+                if (character != '<')
+                    continue;
+
+                // As soon as we find the XML opening tag, break and return the rest of the string, else XElement.Parse fails
+                substringStart = i;
+                break;
+            }
+
+            return inString.Substring(substringStart);
+        }
+
+        public static DateTimeOffset GetPublishDate(XElement item)
+        {
+            var publishDate = item.GetElementValue("pubDate");
+            DateTimeOffset pubDate;
+            try
+            {
+                pubDate = Skybrud.Essentials.Time.TimeUtils.Rfc822ToDateTimeOffset(publishDate);
+            }
+            catch (Exception e)
+            {
+                // special dateformat, try normal C# date parser
+            }
+
+            DateTimeOffset.TryParse(publishDate, out pubDate);
+            return pubDate;
         }
 
     }
